Test coms with xbee

+++ rover-code/embedded/MainBodyBoard/src/DriveBase/Motor.cpp
@@ -2,7 +2,7 @@
 
 Motor::Motor(PWM_PINS pwm_pin){
     this->pwm_pin = pwm_pin;
-    pinMode(this->pwm_pin, OUTPUT);
+    pinMode((int)this->pwm_pin, OUTPUT);
     motor.attach(this->pwm_pin);  // Assuming 'motor' is a member variable of the Motor class
 }
